[#76] Re-structure BackupModal
