+++ universal-login-example/src/components/Modals.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import DevicesModal from '../views/Modals/DevicesModal';
 import PostModal from '../views/Modals/PostModal';
-import BackupModal from '../views/Modals/BackupModal';
+import BackupModal from './BackupModal';
 import PropTypes from 'prop-types';
 
 class Modals extends Component {
@@ -24,17 +24,6 @@ class Modals extends Component {
     this.props.emitter.emit('showModal', null);
   }
 
-  showAccount() {
-    this.hideModal();
-    this.props.emitter.emit('setView', 'Account');
-  }
-
-  async printScreen() {
-    await this.hideModal();
-    await window.print();
-    this.showModal('backup');
-  }
-
   render() {
     switch (this.state.modal) {
     case 'devices':
@@ -42,7 +31,7 @@ class Modals extends Component {
     case 'post':
       return <PostModal hideModal={this.hideModal.bind(this)} />;
     case 'backup':
-      return <BackupModal printScreen={this.printScreen.bind(this)} showAccount={this.showAccount.bind(this)}/>;
+      return <BackupModal emitter={this.props.emitter} />;
     default:
       return null;
     }
